requirements.txt
cryptography
django-jazzmin
django-debug-toolbar
Pillow
requests